backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests

groq

youtube-transcript-api
yt-dlp

openai-whisper
ffmpeg-python
pydub

